fix: logica do PIN
