+++ users/serializers.py
@@ -13,18 +13,8 @@ class EquipeSerializer(serializers.ModelSerializer):
         fields = ['id', 'nome']
 
 class UserSerializer(serializers.ModelSerializer):
-    loja = serializers.PrimaryKeyRelatedField(
-        queryset=Loja.objects.all(),
-        allow_null=True,
-        required=False,
-    )
-    equipe = serializers.PrimaryKeyRelatedField(
-        queryset=Equipe.objects.all(),
-        allow_null=True,
-        required=False,
-    )
-    password = serializers.CharField(write_only=True, required=False)
-    pin = serializers.CharField(write_only=True, required=False, allow_blank=True, allow_null=True)
+    loja = LojaSerializer(read_only=True)
+    equipe = EquipeSerializer(read_only=True)
 
     class Meta:
         model = CustomUser
@@ -38,54 +28,4 @@ class UserSerializer(serializers.ModelSerializer):
             'loja',
             'equipe',
             'is_active',
-            'pin',
-            'password',
         ]
-
-    def to_representation(self, instance):
-        data = super().to_representation(instance)
-        data['loja'] = LojaSerializer(instance.loja).data if instance.loja else None
-        data['equipe'] = EquipeSerializer(instance.equipe).data if instance.equipe else None
-        return data
-
-    def validate_pin(self, value):
-        if value in (None, ''):
-            return None
-        if len(value) != 4 or not value.isdigit():
-            raise serializers.ValidationError('PIN deve ter exatamente 4 digitos numericos.')
-        return value
-
-    def validate(self, attrs):
-        cargo = attrs.get('cargo')
-        if self.instance:
-            cargo = cargo or self.instance.cargo
-
-        pin = attrs.get('pin')
-        if self.instance and pin is None:
-            pin = self.instance.pin
-
-        if cargo == 'VENDEDOR' and not pin:
-            raise serializers.ValidationError({'pin': 'PIN e obrigatorio para vendedores.'})
-
-        if not self.instance and not attrs.get('password'):
-            raise serializers.ValidationError({'password': 'Senha e obrigatoria.'})
-
-        return attrs
-
-    def create(self, validated_data):
-        password = validated_data.pop('password')
-        user = CustomUser(**validated_data)
-        user.set_password(password)
-        user.save()
-        return user
-
-    def update(self, instance, validated_data):
-        password = validated_data.pop('password', None)
-        if password:
-            instance.set_password(password)
-
-        for field, value in validated_data.items():
-            setattr(instance, field, value)
-
-        instance.save()
-        return instance
